Task 1 added

+++ HelloConsole/HelloDiv/Program.cs
@@ -1,10 +1,10 @@
 ﻿Console.WriteLine("Input first digit = ");
 
-int a = Int32.Parse(Console.ReadLine());
+double a = Int32.Parse(Console.ReadLine());
 
 Console.WriteLine("Input second digit = ");
 
-int b = Int32.Parse(Console.ReadLine());
+double b = Int32.Parse(Console.ReadLine());
 
 double divison = a / b;
